Make sure trna file size is more than 0

+++ mag_annotator/annotate_bins.py
@@ -1,7 +1,7 @@
 from skbio.io import read as read_sequence
 from skbio.io import write as write_sequence
 from skbio import Sequence
-from os import path, mkdir
+from os import path, mkdir, stat
 from shutil import rmtree, copy2
 import pandas as pd
 from datetime import datetime
@@ -371,7 +371,7 @@ def run_trna_scan(fasta, output_loc, fasta_name, threads=10, verbose=True):
     raw_trnas = path.join(output_loc, 'raw_trnas.txt')
     run_process(['tRNAscan-SE', '-G', '-o', raw_trnas, '--thread', str(threads), fasta], verbose=verbose)
     processed_trnas = path.join(output_loc, 'trnas.tsv')
-    if path.isfile(raw_trnas):
+    if path.isfile(raw_trnas) and stat(raw_trnas).st_size > 0:
         trna_frame = pd.read_csv(raw_trnas, sep='\t', skiprows=[0, 2], index_col=0)
         trna_frame.insert(0, 'fasta', fasta_name)
         trna_frame.to_csv(processed_trnas, sep='\t')
